separate Controller
for best practice
