+++ templates/trick_image/confirmDeleteTrickImageModal.twig
