+++ templates/trick_image/formTrickImage.twig
@@ -33,7 +33,7 @@
 			    	{% for trickImage in trickImages %}
 			    	<article>
 				    	<hr>
-			    		{% include 'snow/confirmDeleteTrickImageModal.twig' %}
+			    		{% include 'trick_image/confirmDeleteTrickImageModal.twig' %}
 			    		<div class="row">
 			    			<div class="col-1 text-center">
 				    			<em><strong>{{ loop.index }}</strong></em><br>
