+++ templates/trick_video/confirmDeleteTrickVideoModal.twig
