+++ templates/trick_video/formTrickVideo.twig
@@ -34,7 +34,7 @@
 			    	{% for trickVideo in trickVideos %}
 			    	<article>
 				    	<hr>
-			    		{% include 'snow/confirmDeleteTrickVideoModal.twig' %}
+			    		{% include 'trick_video/confirmDeleteTrickVideoModal.twig' %}
 			    		<div class="row">
 			    			<div class="col-1 text-center">
 				    			<em><strong>{{ loop.index }}</strong></em><br>
